requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
